+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stackframe/stack-backend
 
+## 2.5.8
+
+### Patch Changes
+
+- Improved docs
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.8
+  - @stackframe/stack-emails@2.5.8
+
 ## 2.5.7
 
 ### Patch Changes