+++ packages/stack/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @stackframe/stack
 
+## 2.5.8
+
+### Patch Changes
+
+- Improved docs
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.8
+  - @stackframe/stack-ui@2.5.8
+  - @stackframe/stack-sc@2.5.8
+
 ## 2.5.7
 
 ### Patch Changes